chore: use "rule" instead of "lint"
