+++ selene-lib/src/lints/compare_nan.rs
@@ -8,12 +8,12 @@ use full_moon::{
 
 pub struct CompareNanLint;
 
-impl Rule for CompareNanLint {
+impl Lint for CompareNanLint {
     type Config = ();
     type Error = Infallible;
 
     const SEVERITY: Severity = Severity::Error;
-    const RULE_TYPE: RuleType = RuleType::Correctness;
+    const RULE_TYPE: LintType = LintType::Correctness;
 
     fn new(_: Self::Config) -> Result<Self, Self::Error> {
         Ok(CompareNanLint)
