+++ selene-lib/src/lints/constant_table_comparison.rs
@@ -10,12 +10,12 @@ use full_moon::{
 
 pub struct ConstantTableComparisonLint;
 
-impl Rule for ConstantTableComparisonLint {
+impl Lint for ConstantTableComparisonLint {
     type Config = ();
     type Error = Infallible;
 
     const SEVERITY: Severity = Severity::Error;
-    const RULE_TYPE: RuleType = RuleType::Correctness;
+    const RULE_TYPE: LintType = LintType::Correctness;
 
     fn new(_: Self::Config) -> Result<Self, Self::Error> {
         Ok(ConstantTableComparisonLint)
