+++ selene-lib/src/lints/duplicate_keys.rs
@@ -11,12 +11,12 @@ use full_moon::{
 
 pub struct DuplicateKeysLint;
 
-impl Rule for DuplicateKeysLint {
+impl Lint for DuplicateKeysLint {
     type Config = ();
     type Error = Infallible;
 
     const SEVERITY: Severity = Severity::Error;
-    const RULE_TYPE: RuleType = RuleType::Correctness;
+    const RULE_TYPE: LintType = LintType::Correctness;
 
     fn new(_: Self::Config) -> Result<Self, Self::Error> {
         Ok(DuplicateKeysLint)
