+++ selene-lib/src/lints/must_use.rs
@@ -5,12 +5,12 @@ use std::convert::Infallible;
 
 pub struct MustUseLint;
 
-impl Rule for MustUseLint {
+impl Lint for MustUseLint {
     type Config = ();
     type Error = Infallible;
 
     const SEVERITY: Severity = Severity::Warning;
-    const RULE_TYPE: RuleType = RuleType::Correctness;
+    const RULE_TYPE: LintType = LintType::Correctness;
 
     fn new(_: Self::Config) -> Result<Self, Self::Error> {
         Ok(MustUseLint)
