+++ selene-lib/src/lints/undefined_variable.rs
@@ -13,12 +13,12 @@ lazy_static::lazy_static! {
     static ref VARARG_STRING: String = "...".to_owned();
 }
 
-impl Rule for UndefinedVariableLint {
+impl Lint for UndefinedVariableLint {
     type Config = ();
     type Error = Infallible;
 
     const SEVERITY: Severity = Severity::Error;
-    const RULE_TYPE: RuleType = RuleType::Correctness;
+    const RULE_TYPE: LintType = LintType::Correctness;
 
     fn new(_: Self::Config) -> Result<Self, Self::Error> {
         Ok(UndefinedVariableLint)
